+++ spec/nodeApiSpec.js
@@ -30,4 +30,24 @@ describe('js api', () => {
     expect(p.code).toBe(0)
     expect(p.stdout.trim()).toBe('project-watch-handlebars-test1')
   })
+
+  it('mode: string, preserveExtension multiple files', () => {
+    const output = './spec/assets/example/string-mode-multiple-test'
+    shell.rm('-rf', output)
+    const input = './spec/assets/example/*.*'
+    const prev = shell.ls(input)
+    shell.mkdir('-p',output)
+    tool({
+      input,
+      output,
+      mode: 'string',
+      preserveExtension: true
+    })
+    expect(shell.ls(output+'/*.*').length).toBe(prev.length)
+    expect(shell.cat(output+'/data1_working_ts.ts')+'').toContain(`export const data1_working_ts: string = "export interface data1_definition`)
+    expect(shell.exec('npx tsc '+output+'/data1_working_ts.ts').code).toBe(0)
+    const p  =shell.exec(`node -e "console.log(require('${output+'/data1_working_ts.js'}').data1_working_ts)"`)
+    expect(p.code).toBe(0)
+    expect(p.stdout.trim()).toContain('export interface data1_definition')
+  })
 })